# docs/swagger.yaml
basePath: /api/v1/
definitions:
...
  dtos.PostOutput:
    properties:
      author_id:
        type: integer
      author_username:
        type: string
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      modified_at:
        type: string
      score:
        type: integer
      space_id:
        type: integer
      space_username:
        type: string
      title:
        type: string
    type: object
...
  response.ErrorResponse:
    properties:
      code:
        type: string
      errors:
        additionalProperties:
          type: string
        type: object
      msg:
        type: string
      status:
        type: integer
    type: object
  response.SuccessResponse:
    properties:
      code:
        type: string
      data: {}
      msg:
        type: string
      status:
        type: integer
    type: object
info:
  contact: {}
  title: Veteria (Previously Voter)
  version: "1.0"
paths:
...
  /posts:
    get:
      consumes:
      - application/json
      description: Get all posts by page. This URL is for home page.
      parameters:
      - description: 'authorization token (value: Bearer <jwt-token>)'
        in: header
        name: Authorization
        required: true
        type: string
      - description: page number
        in: query
        name: page
        required: true
        type: integer
      - description: \
        in: query
        name: sort_by
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully fetched
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dtos.PostOutput'
                  type: array
              type: object
        "400":
          description: falied
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Access Denied
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      tags:
      - posts
  /posts/:postId:
    delete:
      consumes:
      - application/json
      description: delete a post
      parameters:
      - description: 'authorization token (value: Bearer <jwt-token>)'
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: successfully deleted
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: falied
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      tags:
      - posts
    get:
      consumes:
      - application/json
      description: get a post
      produces:
      - application/json
      responses:
        "200":
          description: successfully fetched
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dtos.PostOutput'
              type: object
        "400":
          description: falied
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      tags:
      - posts
    put:
      consumes:
      - application/json
      description: update a post
      parameters:
      - description: 'authorization token (value: Bearer <jwt-token>)'
        in: header
        name: Authorization
        required: true
        type: string
      - description: post title
        in: body
        name: title
        required: true
        schema:
          type: string
      - description: post content
        in: body
        name: content
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: successfully updated
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: falied
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      tags:
      - posts
...
  /posts/:postId/downvote:
    post:
      consumes:
      - application/json
      description: downvote to a post
      parameters:
      - description: 'authorization token (value: Bearer <jwt-token>)'
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: successfully saved
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: falied
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      tags:
      - posts
  /posts/:postId/upvote:
    post:
      consumes:
      - application/json
      description: upvote to a post
      parameters:
      - description: 'authorization token (value: Bearer <jwt-token>)'
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: successfully saved
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: falied
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      tags:
      - posts
  /posts/:postId/votes:
    delete:
      consumes:
      - application/json
      description: delete vote of a post
      parameters:
      - description: 'authorization token (value: Bearer <jwt-token>)'
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: successfully saved
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: falied
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      tags:
      - posts
  /posts/home:
    get:
      consumes:
      - application/json
      description: Get user home posts. can be sort by "trending", "score", "date".
      parameters:
      - description: 'authorization token (value: Bearer <jwt-token>)'
        in: header
        name: Authorization
        required: true
        type: string
      - description: page number
        in: query
        name: page
        required: true
        type: integer
      - description: \
        in: query
        name: sort_by
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully fetched
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dtos.PostOutput'
                  type: array
              type: object
        "400":
          description: falied
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Access Denied
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      tags:
      - posts
...
  /spaces/:spaceId/posts:
    get:
      consumes:
      - application/json
      description: get all posts of a space by page
      parameters:
      - description: 'authorization token (value: Bearer <jwt-token>)'
        in: header
        name: Authorization
        required: true
        type: string
      - description: page number
        in: query
        name: page
        required: true
        type: integer
      - description: \
        in: query
        name: sort_by
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully fetched
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dtos.PostOutput'
                  type: array
              type: object
        "400":
          description: falied
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Access Denied
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      tags:
      - posts
    post:
      consumes:
      - application/json
      description: create a post
      parameters:
      - description: 'authorization token (value: Bearer <jwt-token>)'
        in: header
        name: Authorization
        required: true
        type: string
      - description: post title
        in: body
        name: title
        required: true
        schema:
          type: string
      - description: post content
        in: body
        name: content
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: successfully created
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dtos.PostOutput'
              type: object
        "400":
          description: falied
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      tags:
      - posts
...
swagger: "2.0"